requirements.txt
# Streamlit
streamlit

# LangChain
python-dotenv
markitdown[all]
langchain
langchain-core
langchain-openai
pydantic